Lots of good stuff.

+++ censorman_v2/src/os/os_core.h
@@ -302,7 +302,7 @@ void   thread_close(Thread thread);
 void   thread_join_many(Thread *threads, s64 thread_count);
 void   thread_close_many(Thread *threads, s64 thread_count);
 
-ThreadValuesRange thread_range(u32 thread_index, u32 thread_count, u32 values_count);
+ThreadValuesRange thread_range(s64 thread_index, s64 thread_count, u64 values_count);
 
 Barrier barrier_create(s64 thread_count);
 b32     barrier_sync(Barrier *barrier);
